@@ -1,7 +1,7 @@
 class AppController {
-  static getHomepage(request, response) {
-    response.status(200).send('Hello Holberton School!');
+  static getHomepage(req, res) {
+    res.status(200).send('Hello Holberton School!');
   }
 }
 
-module.exports = AppController;
+export default AppController;
