@@ -1,6 +1,6 @@
-const express = require('express');
-const AppController = require('../controllers/AppController');
-const StudentsController = require('../controllers/StudentsController');
+import express from 'express';
+import AppController from '../controllers/AppController';
+import StudentsController from '../controllers/StudentsController';
 
 const router = express.Router();
 
@@ -8,4 +8,4 @@ router.get('/', AppController.getHomepage);
 router.get('/students', StudentsController.getAllStudents);
 router.get('/students/:major', StudentsController.getAllStudentsByMajor);
 
-module.exports = router;
+export default router;
