@@ -1,12 +1,10 @@
 process.stdout.write('Welcome to Holberton School, what is your name?\n');
 
-process.stdin.on('readable', () => {
-  const name = process.stdin.read();
-  if (name) {
-    process.stdout.write(`Your name is: ${name}`);
-  }
-});
+process.stdin.setEncoding('utf8');
 
-process.stdin.on('end', () => {
-  process.stdout.write('This important software is now closing\n');
+process.stdin.on('data', (data) => {
+  const name = data.trim();
+  console.log(`Your name is: ${name}`);
+  console.log('This important software is now closing');
+  process.stdin.end();
 });
