@@ -1,38 +1,42 @@
-const fs = require('fs').promises;
+const fs = require('node:fs');
 
-function countStudents(path) {
+function parseStudents(dataLine) {
+  const csStudent = [];
+  let csCount = 0;
+  const sweStudent = [];
+  let sweCount = 0;
+  for (let i = 0; i < dataLine.length; i += 1) {
+    const studentsData = dataLine[i].split(',');
+    if (studentsData[3] === 'CS') {
+      csCount += 1;
+      csStudent.push(studentsData[1]);
+    } else if (studentsData[3] === 'SWE') {
+      sweCount += 1;
+      sweStudent.push(studentsData[1]);
+    }
+  }
+  return {
+    csCount,
+    csStudent,
+    sweCount,
+    sweStudent,
+  };
+}
+
+function countStudents(file) {
   return new Promise((resolve, reject) => {
-    fs.readFile(path, 'utf8')
-      .then((data) => {
-        const lines = data.trim().split('\n');
-        const headers = lines[0].split(',');
-        const fieldIndex = headers.indexOf('field');
-        const firstNameIndex = headers.indexOf('firstname');
-        
-        const students = lines.slice(1).filter(line => line.length > 0);
-        console.log(`Number of students: ${students.length}`);
-        
-        const fields = {};
-        students.forEach((student) => {
-          const row = student.split(',');
-          const field = row[fieldIndex];
-          const firstName = row[firstNameIndex];
-          
-          if (!fields[field]) {
-            fields[field] = [];
-          }
-          fields[field].push(firstName);
-        });
-        
-        for (const [field, names] of Object.entries(fields)) {
-          console.log(`Number of students in ${field}: ${names.length}. List: ${names.join(', ')}`);
-        }
-        
-        resolve();
-      })
-      .catch(() => {
-        reject(new Error('Cannot load the database'));
-      });
+    fs.readFile(file, 'utf8', (err, data) => {
+      if (err) {
+        return reject(new Error('Cannot load the database'));
+      }
+      const dataLine = data.split('\n');
+      if (dataLine.length > 0 && dataLine[dataLine.length - 1].trim() === '') {
+        dataLine.pop();
+      }
+      dataLine.shift();
+      const result = parseStudents(dataLine);
+      return resolve(`Number of students: ${dataLine.length}\nNumber of students in CS: ${result.csCount}. List: ${result.csStudent.join(', ')}\nNumber of students in SWE: ${result.sweCount}. List: ${result.sweStudent.join(', ')}`);
+    });
   });
 }
 
