@@ -1,12 +1,12 @@
 const express = require('express');
 
 const app = express();
-const port = 1245;
+const PORT = 1245;
 
 app.get('/', (req, res) => {
   res.send('Hello Holberton School!');
 });
 
-app.listen(port);
+app.listen(PORT);
 
 module.exports = app;
