@@ -1,11 +1,11 @@
 import express from 'express';
-import routes from './routes';
+import routes from './routes/index';
 
 const app = express();
-const port = 1245;
+const PORT = 1245;
 
 app.use('/', routes);
 
-app.listen(port);
+app.listen(PORT);
 
 export default app;
